backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Data Processing
pandas==2.2.0
numpy==1.26.3
chardet==5.2.0

# Financial Calculations
scipy==1.12.0

# Market Data
yfinance==0.2.36

# ML Libraries
prophet==1.1.5
scikit-learn==1.4.0
xgboost==2.0.3

# API & Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# AI Integration
google-generativeai==0.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
